reafactor: 약간의 리팩토링
